README added - ideas + goals for the project
A readme file was added, detailing the intent of the project, with
examples where it might be useful and it's end goal, to think a step
ahead of the user, so that they may always be prepared. It will also
include resource and library information used as they are used.